# openapi.yaml
openapi: 3.0.0
info:
  title: Remote Inspection Capture Tool API
  version: "1.0.0"
  description: |
    高速リモート検収ツール REST API
    認可は HTTP 401/403 で実装
    全レスポンスは code/message/data を含む統一フォーマット
servers:
  - url: /api
    description: API Gateway HTTP API
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "INVALID_REQUEST"
        message:
          type: string
        details:
          type: object
    SuccessResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: string
          example: "SUCCESS"
        message:
          type: string
        data:
          type: object
    User:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [viewer, capturer, editor, admin]
        createdAt:
          type: string
          format: date-time
    Inspection:
      type: object
      properties:
        inspectionId:
          type: string
        title:
          type: string
        place:
          type: string
        status:
          type: string
          enum: [draft, running, closed]
        realtimeRef:
          type: string
          description: "IVS Stage ID or other video provider reference"
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
...
    GuestSessionToken:
      type: object
      properties:
        guestSessionToken:
          type: string
        expiresIn:
          type: integer
          description: "Expiry time in seconds"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: "認証なし or トークン無効"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
  /me:
    get:
      tags: [Auth]
      summary: "ログインユーザー情報"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /inspections:
    post:
      tags: [Inspections]
      summary: "新規案件作成"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, place]
              properties:
                title:
                  type: string
                place:
                  type: string
      responses:
        "201":
          description: "作成成功"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Inspection"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /inspections/{inspectionId}:
    get:
      tags: [Inspections]
      summary: "案件詳細取得"
      security:
        - bearerAuth: []
      parameters:
        - name: inspectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Inspection"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: "案件なし"
...
  /public/viewer-sessions:
    post:
      tags: [Public]
      summary: "ゲスト session token 発行 (ログイン不要)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [viewerAccessToken]
              properties:
                viewerAccessToken:
                  type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/GuestSessionToken"
        "400":
          description: "Invalid or expired token"
        "404":
          description: "Token not found"
...